% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahs_get_data.R
\name{ahs_get_data}
\alias{ahs_get_data}
\title{Get AHSQC data}
\usage{
ahs_get_data(date = NULL, dir = NULL)
}
\arguments{
\item{date}{NULL or character string of a date in the format "YYYY-MM-DD"}

\item{dir}{NULL or character string with file path to ahs-data folder.}
}
\description{
Retrieves the most recent download of AHSQC data or the download from a specific date.
}
\details{
If \code{date} = NULL, then the function will return the most recently downloaded dataset from the ahs-data folder. 

If \code{dir} = NULL, the function will assume that the command is being run in a working directory of the following type: ".../alfresco/Sites/[SITE NAME]/documentLibrary/code".

If date is a character string of a date, then the function retuns the data downloaded on that date, if it exists.  If the date does not exist in the ahs-data folder, the function throws an error.
}
\examples{
# Not run:
# d0 <- ahs_get_data()
}
\keyword{AHSQC}
