% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahsqc_label.R
\name{ahsqc_label}
\alias{ahsqc_label}
\title{Creates a label or shows label}
\usage{
ahsqc_label(x, value = NULL)
}
\arguments{
\item{x}{data vector}

\item{value}{string, label value. If left NULL, returns label}
}
\description{
Creates a label or shows label
}
\details{
Creates an empty data frame of nrows and ncols
}
\examples{
# Not run:
}
\keyword{AHSQC}
