% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_entry_all.R
\name{binary_entry_all}
\alias{binary_entry_all}
\title{Creates a binary entry for a table}
\usage{
binary_entry_all(out, x, dt, xlab = NULL, level = c("Yes", "1"),
  pvalue = TRUE, fmt = "norm_fmt", pvalue_fmt = function(x,
  test_method) {     formatp(x, digits = 3) \%|\% "<sup>" \%|\% test_method
  \%|\% "</sup>" })
}
\arguments{
\item{out}{empty vector}

\item{x}{variable for row in table}

\item{dt}{data table}

\item{xlab}{Label for entry in table}

\item{pvalue}{Boolean}

\item{fmt}{Two values are "norm_fmt" or "count_fmt"}

\item{pvalue_fmt}{A function that takes a pvalue and a test method}
}
\description{
Creates a binary entry for a table
}
\details{

}
\examples{
# Not run:
}
\keyword{AHSQC}
