% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_approx.R
\name{chi_approx}
\alias{chi_approx}
\title{Muffles warning for `chisq.test`}
\usage{
chi_approx(w)
}
\arguments{
\item{w}{character string that may contain "Chi-squared approximation may be incorrect"}
}
\description{
Muffles warning for `chisq.test`
}
\details{
Use as argument for warning in `chisq.test` to muffle warning of "Chi-squared approximation may be incorrect"
}
\examples{
# Not run:
data(iris)
Mat <- table(iris$Species == "setosa", iris$Petal.Width > 1)
chisq.test(Mat, correct = FALSE), warning = chi_approx)
}
\keyword{AHSQC}
