% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_entry_all.R
\name{cont_entry_all}
\alias{cont_entry_all}
\title{Creates a continuous entry for a table}
\usage{
cont_entry_all(out, x, dt, xlab = NULL, pvalue = TRUE,
  pvalue_fmt = function(x, test_method) {     formatp(x, digits = 3)
  \%|\% "<sup>" \%|\% test_method \%|\% "</sup>" })
}
\arguments{
\item{out}{empty vector}

\item{x}{variable for row in table}

\item{dt}{data table}

\item{xlab}{Label for entry in table}

\item{pvalue}{Boolean}

\item{pvalue_fmt}{A function that takes a pvalue and a test method}
}
\description{
Creates a continuous entry for a table
}
\details{

}
\examples{
# Not run:
}
\keyword{AHSQC}
