% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_entry.R
\name{empty_entry}
\alias{empty_entry}
\title{Creates an empty entry in table}
\usage{
empty_entry(out, y, dt, fill = "")
}
\arguments{
\item{out}{empty vector}

\item{y}{grouping variable}

\item{dt}{data table}

\item{fill}{string}
}
\description{
Creates an empty entry in table
}
\details{

}
\examples{
# Not run:
}
\keyword{AHSQC}
