% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_entry_all.R
\name{empty_entry_all}
\alias{empty_entry_all}
\title{Creates an empty entry in table}
\usage{
empty_entry_all(out, dt, fill = "")
}
\arguments{
\item{out}{empty vector}

\item{dt}{data table}

\item{fill}{string}
}
\description{
Creates an empty entry in table
}
\details{

}
\examples{
# Not run:
}
\keyword{AHSQC}
