% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_pvalue.R
\name{empty_pvalue}
\alias{empty_pvalue}
\title{Returns empty string for pvalue}
\usage{
empty_pvalue(x, test_method)
}
\arguments{
\item{x}{A p-value}

\item{test_method}{A string, statistical test name}
}
\description{
Returns empty string for pvalue
}
\details{
returns an empty string. Used for pvalue_fmt argument in binary_entry, cat_entry, etc.
}
\examples{
# Not run:
data(iris)
cat_entry(c(), x = species, y = ___, pvalue_fmt = empty_pvalue)
}
\keyword{AHSQC}
