% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatpct.R
\name{formatpct}
\alias{formatpct}
\title{Format 2x2 table for html}
\usage{
formatpct(M, fmt = "\%1.0f (\%s)\%s")
}
\arguments{
\item{M}{a 2x2 matrix}

\item{fmt}{string that represents a format}
}
\description{
Format 2x2 table for html
}
\details{

}
\examples{
# Not run:
data(iris)
Mat <- table(iris$Species == "setosa", iris$Petal.Width > 1)
formatpct(M = mat)
}
\keyword{AHSQC}
