% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garbage_pvalue.R
\name{garbage_pvalue}
\alias{garbage_pvalue}
\title{Returns pvalue format with NV specification}
\usage{
garbage_pvalue(x, test_method)
}
\arguments{
\item{x}{A p-value}

\item{test_method}{A string, statistical test name}
}
\description{
Returns pvalue format with NV specification
}
\details{
Returns a specific format. Used for pvalue_fmt argument in binary_entry, cat_entry, etc. 
NV	indicates: As the baseline demographic and pre-operative characteristics indicate, the patients in each repair group differ in terms of baseline disease severity and other characteristics associated with repair outcomes. In light of the imbalance, an unadjusted comparison between groups would not be interpretable as it would be confounded by factors related the surgeon’s choice of surgical approach. For this reason, post operative outcomes were not formally compared between groups in the unadjusted analysis.
}
\examples{
# Not run:
data(iris)
cat_entry(c(), x = species, y = ___, pvalue_fmt = garbage_pvalue)
}
\keyword{AHSQC}
