% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_standard_tables.R
\name{generate_standard_tables}
\alias{generate_standard_tables}
\title{Generate standard AHSQC tables}
\usage{
generate_standard_tables(changes = FALSE, y, data)
}
\arguments{
\item{changes}{a boolean. Set to TRUE to indicate that changes have been made to at least one table.}

\item{y}{a character string to indicate group variable for tables.}

\item{data}{a data table.}
}
\description{
Generates standard AHSQC tables in HTML for .Rmd files. \code{generate_standard_tables}
will print standard 
tables with no changes by default. If changes to individual tables are desired, they 
can be made with \code{get_standard_table} using separate .R files in the working directory.
}
\details{
Function returns the standard set of tables for the AHSQC in HTML code.

If \code{changes = TRUE}, the code for each table should be in individual .R files 
with the \code{tbl}n\code{.R} nomenclature (ex. tbl1.R, tbl2.R, etc.) should be stored in the 
working directory.  The function \code{get_standard_table} will do this.
}
\examples{
# Not run:
# d0 <- ahs_get_data()
# d1 <- d0[["analyticview"]]
# generate_standard_tables(dt = d1, y = "e_gender")  
}
\keyword{AHSQC}
