% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_standard_tables_all.R
\name{generate_standard_tables_all}
\alias{generate_standard_tables_all}
\title{Generate standard AHSQC tables}
\usage{
generate_standard_tables_all(changes = FALSE, data, date,
  format = "rmd", module = "ventral", display = TRUE)
}
\arguments{
\item{changes}{a boolean. Set to TRUE to indicate that changes have been made to at least one table.}

\item{data}{a data table.}

\item{format}{a string that describes the output format. Currently only "shiny" or
"rmd". Default is "rmd"}

\item{module}{a string that indicates which data module, ventral or inguinal, is used to create
tables. Set to "ventral" for ventral module and "inguinal" for inguinal table. Once cannot
generate tables for ventral and inguinal modules together.}

\item{out}{empty vector}

\item{x}{variable for row in table}

\item{dt}{data table}

\item{xlab}{Label for entry in table}

\item{out}{empty vector}

\item{x}{variable for row in table}

\item{dt}{data table}

\item{xlab}{Label for entry in table}

\item{pvalue}{Boolean}

\item{fmt}{Two values are "norm_fmt" or "count_fmt"}

\item{pvalue_fmt}{A function that takes a pvalue and a test method}

\item{out}{empty vector}

\item{dt}{data table}

\item{fill}{string}
}
\description{
Generates standard AHSQC tables in HTML for .Rmd files. \code{generate_standard_tables}
will print standard
tables with no changes by default. If changes to individual tables are desired, they
can be made with \code{get_standard_table} using separate .R files in the working directory.
}
\details{
Function returns the standard set of tables for the AHSQC in HTML code.

If \code{changes = TRUE}, the code for each table should be in individual .R files
with the \code{tbl}n\code{.R} nomenclature (ex. tbl1.R, tbl2.R, etc.) should be stored in the
working directory.  The function \code{get_standard_table} will do this.






}
\examples{
# Not run:
# d0 <- ahs_get_data()
# d1 <- d0[["analyticview"]]
# generate_standard_tables(dt = d1, y = "e_gender")
Creates an entry with number of unique observations

# Not run:
Creates a categorical entry for a table

# Not run:
Creates an empty entry in table

# Not run:
}
\keyword{AHSQC}
