% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ORACLE_formula.R
\name{get_ORACLE_formula}
\alias{get_ORACLE_formula}
\title{Get ORACLE formula}
\usage{
get_ORACLE_formula(y)
}
\arguments{
\item{y}{Dependent variable}
}
\description{
Returns the ORACLE regression formula
}
\details{
Returns the formula used in the ORACLE regression models
}
\examples{
# Not run:
# d0 <- ahs_get_data()
}
\keyword{AHSQC,}
\keyword{ORACLE}
