% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_out.R
\name{get_out}
\alias{get_out}
\title{Creates an empty data frame}
\usage{
get_out(nrow, ncol)
}
\arguments{
\item{nrow}{integer, number of rows}

\item{ncol}{integer, number of columns}
}
\description{
Creates an empty data frame
}
\details{
Creates an empty data frame of nrows and ncols
}
\examples{
# Not run:
get_out(3,4)
}
\keyword{AHSQC}
