% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_table_all2.R
\name{get_standard_table_all2}
\alias{get_standard_table_all2}
\title{Get an individual standard table for AHSQC}
\usage{
get_standard_table_all2(tbl = NULL, data, print = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{tbl}{NULL or integer (1-9) to identify which table to return.}

\item{data}{a data table}

\item{print}{a boolean. If TRUE, the code for specified table will be returned
in a tet string. If FALSE, the code for specified table will be written to a
separate .R file in the working directory with \code{tbl}n\code{.R} nomenclature.}

\item{overwrite}{a boolean. If a .R file already exists for the table specified and set to TRUE, the file will be overwritten with original code.}
}
\description{
Retrieves the code for the standard AHSQC table specified and returns it in either
a separate file or through a text string.
}
\details{
If \code{overwrite} = FALSE and a .R file exists for the specified table, the function will throw an error.

If \code{print} = TRUE, the \code{overwrite} argument will be ignored.
}
\examples{
# Not run:
# d0 <- ahs_get_data()
# d1 <- d0[["analyticview]]
# get_table(tbl = 1, d1, print = TRUE) ## will print code as text
# get_table(tbl = 9, d1, print = FALSE) ## will write tbl9.R
# get_table(tbl = 9, d1, print = FALSE, overwrite = TRUE) ## will overwrite tbl9.R
}
\keyword{AHSQC}
