% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique_all.R
\name{n_unique_all}
\alias{n_unique_all}
\title{Creates an entry with number of unique observations}
\usage{
n_unique_all(out, x, dt, xlab = NULL)
}
\arguments{
\item{out}{empty vector}

\item{x}{variable for row in table}

\item{dt}{data table}

\item{xlab}{Label for entry in table}
}
\description{
Creates an entry with number of unique observations
}
\details{

}
\examples{
# Not run:
}
\keyword{AHSQC}
