% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{nin}
\alias{nin}
\title{%nin%}
\usage{
nin(x, table)
}
\arguments{
\item{x}{vector: the values to be matched}

\item{y}{vector, list, or data frame: the value to be matched against}
}
\value{
boolean
}
\description{
Opposite of %in%
}
\examples{
# Not run:


}
\keyword{render}
