% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_identifier.R
\name{rand_identifier}
\alias{rand_identifier}
\title{Map an ID variable to randomly generated IDs}
\usage{
rand_identifier(x, seed)
}
\arguments{
\item{x}{A numeric or character vector; the ID variable}

\item{seed}{Seed for random sampling}
}
\description{
Generates a random ID variable based on a previously existing ID variable.
}
\details{
Maps ID values to a random ordering of integers to generate the new ID.
}
\examples{
rand_identifier(LETTERS, 234)
rand_identifier(1:23, 345)
}
\keyword{rand_identifier}
