% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadj_or.R
\name{unadj_or}
\alias{unadj_or}
\title{Returns odds ratio of two variables from data frame}
\usage{
unadj_or(v1, v2, d1)
}
\arguments{
\item{v1}{Character, variable 1 name}

\item{v2}{Character, variable 2 name}

\item{data}{dataframe or datatable}
}
\description{
Returns odds ratio of two variables from data frame
}
\details{
Returns a dataframe with variable, estimate, and CI of odds ratio
}
\examples{
# Not run:
# unadj_or("flg_cmp_postop_ssi", "e_gender", data = d1)
}
\keyword{AHSQC,}
\keyword{OR}
\keyword{odds}
\keyword{ratio,}
